app/requirements.txt
clickhouse-connect==0.10.0
python-dotenv==0.21.0
pydantic==1.9.0
